feat: Add cloudform type for type safety, move validation to provider-cloudformation, validation
