+++ packages/amplify-cli-core/src/customPoliciesUtils.ts
@@ -1,3 +1,7 @@
+export type CustomIAMPolicies = {
+    policies:  CustomIAMPolicy[];
+}
+
 export type CustomIAMPolicy = {
     Action: string[];
     Effect: string;
@@ -6,15 +10,25 @@ export type CustomIAMPolicy = {
 
 
 export const CustomIAMPolicySchema = {
-        type: "object",
-        properties: {
-            Action: {type: "array", items: {type: "string"}, minItems: 1, nullable: false},
-            Effect: {type: "string", items: {type: "string"}, minItems: 1, nullable: false, default: "Allow"},
-            Resource: {type: "array", items: {type: "string"}, minItems: 1, nullable: false},
-        },
-        required: ["Resource", "Action"],
-        additionalProperties: false
+    type: "object",
+    properties: {
+        policies: {
+            type: "array",
+            minItems: 1,
+            items: {
+                type: "object",
+                properties: {
+                    "Action": { type: "array", items: { type: "string" }, minItems: 1, nullable: false },
+                    "Effect": { type: "string",  nullable: false, default: "Allow" },
+                    "Resource": { type: "array", items: { type: "string" }, minItems: 1, nullable: false },
+                },
+                required: ["Resource", "Action"],
+                additionalProperties: false
+            },
+        }
+    }
 }
+
