+++ packages/amplify-provider-awscloudformation/src/pre-push-cfn-processor/cfn-pre-processor.ts
@@ -1,9 +1,10 @@
-import { stateManager, pathManager, readCFNTemplate, writeCFNTemplate, CustomIAMPolicy, CustomPolicyFileContentConstant } from 'amplify-cli-core';
+import { stateManager, pathManager, readCFNTemplate, writeCFNTemplate, CustomIAMPolicy, CustomIAMPolicySchema, CustomIAMPolicies, customExecutionPolicyForFunction, customExecutionPolicyForContainer } from 'amplify-cli-core';
 import * as path from 'path';
 import { ProviderName as providerName } from '../constants';
 import { prePushCfnTemplateModifier } from './pre-push-cfn-modifier';
 import { Fn, Template } from 'cloudform-types';
-import { Printer, AmplifyPrinter } from 'amplify-prompts';
+import { AmplifyPrinter, printer, Printer } from 'amplify-prompts';
+import Ajv from "ajv";
 
 
 const buildDir = 'build';
@@ -40,9 +41,9 @@ export async function writeCustomPoliciesToCFNTemplate(
 ) {
   const { templateFormat, cfnTemplate } = await readCFNTemplate(path.join(resourceDir, cfnFile));
   const customPolicies = stateManager.getCustomPolicies(service, category, resourceName);
-  if (!customPolicies) return;
+  if (!customPolicies || !this.validateCustomPolicies(customPolicies)) return;
 
-  await addCustomPoliciesToCFNTemplate(service, customPolicies, cfnTemplate, filePath, resourceName, {templateFormat} );
+  await addCustomPoliciesToCFNTemplate(service, customPolicies.policies, cfnTemplate, filePath, resourceName, {templateFormat} );
 
 }
 
@@ -57,19 +58,21 @@ export async function addCustomPoliciesToCFNTemplate(
   {templateFormat}: any
   ) {
   let customExecutionPolicy;
-
+  const { envName } = stateManager.getLocalEnvInfo();
   if(service === 'Lambda') {
-    customExecutionPolicy = CustomPolicyFileContentConstant.customExecutionPolicyForFunction;
+    customExecutionPolicy = customExecutionPolicyForFunction;
   }
   if (service === 'ElasticContainer') {
     const roleName = cfnTemplate.Resources.TaskDefinition.Properties.ExecutionRoleArn["Fn::GetAtt"][0];
-    customExecutionPolicy = CustomPolicyFileContentConstant.customExecutionPolicyForContainer;
+    customExecutionPolicy = customExecutionPolicyForContainer;
     const role = Fn.Ref(roleName);
     customExecutionPolicy.Properties.Roles.push(role);
   }
 
   for (const customPolicy of customPolicies) {
-    customExecutionPolicy.Properties.PolicyDocument.Statement.push(customPolicy);
+    const policyWithEnv = await replaceEnvForCustomPolicies(customPolicy, envName);
+    await validateRegexCustomPolicy(customPolicy, resourceName);
+    customExecutionPolicy.Properties.PolicyDocument.Statement.push(policyWithEnv);
   }
 
   if (service === 'Lambda') {
@@ -80,4 +83,76 @@ export async function addCustomPoliciesToCFNTemplate(
   }
 
   await writeCFNTemplate(cfnTemplate, filePath, { templateFormat });
+}
+
+
+//validate the format of actions and ARNs for custom IAM policies
+export async function validateRegexCustomPolicy(customPolicy: CustomIAMPolicy, resourceName: string) : Promise<void> {
+  const resources = customPolicy.Resource;
+  const actions = customPolicy.Action;
+  let resourceRegex = new RegExp('(arn:(aws[a-zA-Z0-9-]*):([a-zA-Z0-9\\-])+:([a-z]{2}(-gov)?-[a-z]+-\\d{1})?:(\\d{12})?:(.*))*');
+  let actionRegex = new RegExp('([a-z0-9])*:([a-z|A-Z|0-9|*]+)*');
+  let wrongResourcesRegex = [];
+  let wrongActionsRegex = [];
+  let errorMessage = "";
+  const printer: Printer = new AmplifyPrinter()
+
+  for (const resource of resources) {
+    if (!resourceRegex.test(resource)) {
+      wrongResourcesRegex.push(resource);
+    }
+  }
+
+  if(resources.includes('*')) {
+    printer.warn(`Warning: You've specified "*" as a custom IAM policy for your ${resourceName}. 
+    This will give your ${resourceName} access to ALL resources in the AWS Account.`)
+  }
+
+  for (const action of actions) {
+    if (!actionRegex.test(action)) {
+      wrongActionsRegex.push(action);
+    }
+  }
+
+  if (wrongResourcesRegex.length > 0) {
+    errorMessage += `\nInvalid ARN format for custom IAM policies in ${resourceName}:\n${wrongResourcesRegex.toString()}\n`;
+  }
+  if (wrongActionsRegex.length > 0) {
+    errorMessage += `\nInvalid actions format for custom IAM policies in ${resourceName}:\n${wrongActionsRegex.toString()}\n`;
+  }
+
+  if (errorMessage.length > 0) {
+    printer.error(errorMessage);
+  }
+}
+
+
+//replace or add env parameter in the front of the resource customers enter to the current env
+export async function replaceEnvForCustomPolicies(policy: CustomIAMPolicy, currentEnv: string) : Promise<CustomIAMPolicy> {
+  let resourceWithEnv = [];
+  let action = policy.Action;
+  let effect = policy.Effect;
+  let customIAMpolicy: CustomIAMPolicy = {
+    Action: action,
+    Effect: effect,
+    Resource: []
+  }
+  for (let resource of policy.Resource){
+    resourceWithEnv.push(resource.replace( '{env}', currentEnv));
+  }
+  customIAMpolicy.Resource = resourceWithEnv;
+  return customIAMpolicy;
+}
+
+
+export async function validateCustomPolicies(data: CustomIAMPolicies, categoryName: string, resourceName: string) : Promise<Boolean> {
+  const ajv = new Ajv();
+  //validate if the policies match the custom IAM policies schema, if not, then not write into the CFN template
+  const validatePolicy = ajv.compile(CustomIAMPolicySchema);
+  if(!validatePolicy(data)) {
+    printer.warn(`The format of custom IAM policies in the ${categoryName} ${resourceName} is not valid`);
+    return false;
+  }
+  //the policies without env will be carried over and merge into the CFN template
+  return true;
 }
